Separated TPM marshalling and kmyth formatting tools (#132)
* Separated TPM marshalling and kmyth formatting tools

* Update marshalling_tools.h

Trivial comment/documentation edit
