+++ src/util/formatting_tools.c
@@ -1,0 +1,297 @@
+/**
+ * formatting_tools.c:
+ *
+ * C library containing data formatting utilities supporting Kmyth
+ */
+
+#include "util/formatting_tools.h"
+
+#include <string.h>
+
+#include <openssl/bio.h>
+#include <openssl/buffer.h>
+#include <openssl/evp.h>
+#include <tss2/tss2_mu.h>
+
+#include "defines.h"
+
+void free_ski(Ski * ski)
+{
+  free(ski->enc_data);
+  ski->enc_data = NULL;
+  ski->enc_data_size = 0;
+}
+
+Ski get_default_ski(void)
+{
+  Ski ret = {
+    .pcr_list = {.count = 0,},
+    .sk_pub = {.size = 0,},
+    .sk_priv = {.size = 0,},
+    .wk_pub = {.size = 0},
+    .wk_priv = {.size = 0},
+    .enc_data = NULL,
+    .enc_data_size = 0
+  };
+  return (ret);
+
+}
+
+//############################################################################
+// get_ski_block_bytes()
+//############################################################################
+int get_ski_block_bytes(char **contents,
+                        size_t * remaining,
+                        uint8_t ** block, size_t * blocksize,
+                        char *delim, size_t delim_len,
+                        char *next_delim, size_t next_delim_len)
+{
+  // check that next (current) block begins with expected delimiter
+  if (strncmp(*contents, delim, delim_len))
+  {
+    kmyth_log(LOG_ERR, "unexpected delimiter ... exiting");
+    return 1;
+  }
+  *contents += delim_len;
+  (*remaining) -= delim_len;
+
+  // find the end of the block 
+  size_t size = 0;
+
+  if (next_delim_len > *remaining)
+  {
+    kmyth_log(LOG_ERR, "unexpectedly reached end of .ski file ... exiting");
+    return 1;
+  }
+  while (strncmp(*contents + size, next_delim, next_delim_len))
+  {
+    size++;
+    if (size + next_delim_len > *remaining)
+    {
+      kmyth_log(LOG_ERR, "unexpectedly reached end of .ski file ... exiting");
+      return 1;
+    }
+  }
+
+  // check that the block is not empty
+  if (size == 0)
+  {
+    kmyth_log(LOG_ERR, "empty .ski block ... exiting");
+    return 1;
+  }
+
+  else
+  {
+    // allocate enough memory for output parameter to hold parsed block data
+    //   - must be allocated here because size is calculated here
+    //   - must be freed by caller because data must be passed back
+    *block = (uint8_t *) malloc(size);
+    if (*block == NULL)
+    {
+      kmyth_log(LOG_ERR, "malloc (%d bytes) error ... exiting", size);
+      return 1;
+    }
+
+    // update output parameters before exiting
+    //   - *block      : block data (for block just parsed)
+    //   - *blocksize  : block data size (for block just parsed)
+    //   - *contents   : pointer to start of next block in .ski file buffer
+    //   - *remaining  : count of bytes yet to be parsed in .ski file buffer
+    memcpy(*block, *contents, size);
+    *blocksize = size;
+    *contents += size;
+    *remaining -= size;
+  }
+
+  return 0;
+}
+
+//############################################################################
+// encodeBase64Data()
+//############################################################################
+int encodeBase64Data(uint8_t * raw_data,
+                     size_t raw_data_size,
+                     uint8_t ** base64_data, size_t * base64_data_size)
+{
+  // check that there is actually data to encode, return error if not
+  if (raw_data == NULL || raw_data_size == 0)
+  {
+    kmyth_log(LOG_ERR, "no input data ... exiting");
+    return 1;
+  }
+
+  BIO *bio_mem = NULL;
+  BIO *bio64 = NULL;
+  BUF_MEM *bioptr = NULL;
+
+  // create a base64 encoding filter BIO
+  if ((bio64 = BIO_new(BIO_f_base64())) == NULL)
+  {
+    kmyth_log(LOG_ERR, "create base64 filter BIO error ... exiting");
+    return 1;
+  }
+
+  // create a 'sink' BIO to write to memory
+  if ((bio_mem = BIO_new(BIO_s_mem())) == NULL)
+  {
+    kmyth_log(LOG_ERR, "create read/write memory sink BIO error" "... exiting");
+    BIO_free_all(bio64);
+    return 1;
+  }
+
+  // assemble the BIO chain in the order bio64 -> bio_mem
+  bio64 = BIO_push(bio64, bio_mem);
+
+  // write the input 'raw data' to the BIO chain
+  if (BIO_write(bio64, raw_data, raw_data_size) != raw_data_size)
+  {
+    kmyth_log(LOG_ERR, "BIO_write() error ... exiting");
+    BIO_free_all(bio64);
+    return 1;
+  }
+
+  // ensure all written data is flushed all the way through chain
+  if (BIO_flush(bio64) != 1)
+  {
+    kmyth_log(LOG_ERR, "BIO_flush() error ... exiting");
+    BIO_free_all(bio64);
+    return 1;
+  }
+
+  // compute memory size of encoded data
+  BIO_get_mem_ptr(bio64, &bioptr);
+  if (bioptr == NULL)
+  {
+    kmyth_log(LOG_ERR, "no underlying BIO_MEM structure ... exiting");
+    BIO_free_all(bio64);
+    return 1;
+  }
+  *base64_data_size = bioptr->length;
+
+  // allocate memory for 'base64_data' output parameter
+  //   - memory allocated here because the encoded data size is known here
+  //   - memory must be freed by the caller because the data passed back
+  *base64_data = (uint8_t *) malloc(*base64_data_size + 1);
+  if (*base64_data == NULL)
+  {
+    kmyth_log(LOG_ERR, "malloc error (%lu bytes) ... exiting",
+              base64_data_size);
+    BIO_free_all(bio64);
+    return 1;
+  }
+
+  // copy encoded data to output parameter and terminate with newline and
+  // null terminator
+  memcpy(*base64_data, bioptr->data, (*base64_data_size) - 1);
+  (*base64_data)[(*base64_data_size) - 1] = '\n';
+  (*base64_data)[(*base64_data_size)] = '\0';
+  kmyth_log(LOG_DEBUG, "encoded %lu bytes into %lu base-64 symbols",
+            raw_data_size, *base64_data_size - 1);
+  // clean-up
+  BIO_free_all(bio64);
+  return 0;
+}
+
+//############################################################################
+// decodeBase64Data()
+//############################################################################
+int decodeBase64Data(uint8_t * base64_data,
+                     size_t base64_data_size,
+                     uint8_t ** raw_data, size_t * raw_data_size)
+{
+  // check that there is actually data to decode, return error if not
+  if (base64_data == NULL || base64_data_size == 0)
+  {
+    kmyth_log(LOG_ERR, "no input data ... exiting");
+    return 1;
+  }
+
+  // check that size of input doesn't exceed limits, return error if it does
+  if (base64_data_size > INT_MAX)
+  {
+    kmyth_log(LOG_ERR,
+              "encoded data length (%lu bytes) > max (%d bytes) ... exiting",
+              base64_data_size, INT_MAX);
+    return 1;
+  }
+
+  // allocate memory for decoded result - size of encoded input is worst case
+  *raw_data = (uint8_t *) malloc(base64_data_size);
+  if (*raw_data == NULL)
+  {
+    kmyth_log(LOG_ERR, "malloc error (%lu bytes) for b64 decode ... exiting",
+              base64_data_size);
+    return 1;
+  }
+
+  // create a base64 decoding filter BIO
+  BIO *bio64 = NULL;
+
+  if ((bio64 = BIO_new(BIO_f_base64())) == NULL)
+  {
+    kmyth_log(LOG_ERR, "create base64 filter BIO error ... exiting");
+    return 1;
+  }
+
+  // create a 'source' BIO to read from memory
+  BIO *bio_mem = NULL;
+
+  if ((bio_mem = BIO_new_mem_buf(base64_data, base64_data_size)) == NULL)
+  {
+    kmyth_log(LOG_ERR, "create source BIO error ... exiting");
+    BIO_free_all(bio64);
+    return 1;
+  }
+
+  // assemble the BIO chain to base64 decode data read from memory
+  bio64 = BIO_push(bio64, bio_mem);
+  // read encoded data through chain, into 'raw_data' decoded output parameter
+  // and terminate with newline
+  int bytes_read = BIO_read(bio64, *raw_data, base64_data_size);
+
+  if (bytes_read < 0)
+  {
+    kmyth_log(LOG_ERR, "error reading bytes from BIO chain ... exiting");
+    BIO_free_all(bio64);
+    return 1;
+  }
+
+  (*raw_data)[bytes_read] = '\0';
+  *raw_data_size = bytes_read;
+  // clean-up
+  BIO_free_all(bio64);
+  return 0;
+}
+
+//############################################################################
+// concat()
+//############################################################################
+int concat(uint8_t ** dest, size_t * dest_length, uint8_t * input,
+           size_t input_length)
+{
+  if (input == NULL || input_length == 0) //nothing to concat
+  {
+    return (0);
+  }
+
+  uint8_t *new_dest = NULL;
+  size_t new_dest_len = *dest_length + input_length;
+  size_t offset = *dest_length;
+
+  if (new_dest_len < *dest_length)  //if we have an overflow
+  {
+    kmyth_log(LOG_ERR, "Maximum array size exceeded ... exiting");
+    return (1);
+  }
+
+  if ((new_dest = realloc(*dest, new_dest_len)) == NULL)
+  {
+    kmyth_log(LOG_ERR, "Ran out of memory ... exiting");
+    return (1);
+  }
+
+  memcpy(&new_dest[offset], input, input_length);
+  *dest = new_dest;
+  *dest_length = new_dest_len;
+  return (0);
+}
